+++ backend/resources/js/Pages/admin/Dashboard.jsx
@@ -5,6 +5,7 @@ import { Badge, Dropdown, StatusPill } from '@openmes/ui';
 import { DataTable } from '@openmes/ui/table';
 import { useDashboardShapes } from '../../lib/useDashboardShapes';
 import { useHotShapes } from '../../components/LiveShapesProvider';
+import OeeGauge, { oeeColor } from '../../components/OeeGauge';
 import AppLayout from '../../layouts/AppLayout';
 import { __ } from '../../lib/i18n';
 import { formatDateTime } from '../../lib/i18n';
@@ -320,17 +321,7 @@ function OeeOverview({ records, lines }) {
     );
 }
 
-// OEE banding — mirrors backend/app/Support/OeeBand.php (red < 65 ≤ yellow < 85 ≤ green).
-const OEE_RED_BELOW = 65;
-const OEE_GREEN_AT_LEAST = 85;
-
-function oeeColor(value) {
-    if (value == null) return 'gray';
-    if (value >= OEE_GREEN_AT_LEAST) return 'green';
-    if (value >= OEE_RED_BELOW) return 'yellow';
-    return 'red';
-}
-
+// Card background tint by OEE band — mirrors backend/app/Support/OeeBand.php.
 function oeeCardClass(value) {
     return {
         green: 'border-om-line bg-om-running-bg',
@@ -338,60 +329,6 @@ function oeeCardClass(value) {
         red: 'border-om-line bg-om-blocked-bg',
         gray: 'border-om-line bg-om-panel',
     }[oeeColor(value)];
-}
-
-function oeeTextClass(value) {
-    return {
-        green: 'text-om-running',
-        yellow: 'text-om-downtime',
-        red: 'text-om-blocked',
-        gray: 'text-om-faint',
-    }[oeeColor(value)];
-}
-
-/**
- * Semicircle OEE gauge with red/yellow/green zones and a needle — React port of
- * the `<x-oee-gauge>` Blade component. Points lie on a unit semicircle centered
- * at (50,50), r=40: p=0 → (10,50), p=50 → (50,10), p=100 → (90,50).
- */
-function OeeGauge({ value, size = 104 }) {
-    const hasValue = value != null;
-    const p = hasValue ? Math.max(0, Math.min(100, Number(value))) : 0;
-    const pointAt = (q, r = 40) => {
-        const a = (q / 100) * Math.PI;
-        return [50 - r * Math.cos(a), 50 - r * Math.sin(a)];
-    };
-    const [rEndX, rEndY] = pointAt(OEE_RED_BELOW);
-    const [yEndX, yEndY] = pointAt(OEE_GREEN_AT_LEAST);
-    const [gEndX, gEndY] = pointAt(100);
-    const [needleX, needleY] = pointAt(p, 35);
-
-    return (
-        <div className="inline-flex flex-col items-center" style={{ width: size }}>
-            <svg viewBox="0 0 100 60" className="w-full h-auto" aria-hidden="true">
-                <path d="M 10 50 A 40 40 0 0 1 90 50" fill="none" stroke="currentColor" strokeWidth="10" className="text-om-line2" />
-                <path d={`M 10 50 A 40 40 0 0 1 ${rEndX} ${rEndY}`} fill="none" stroke="#D6442F" strokeWidth="10" />
-                <path d={`M ${rEndX} ${rEndY} A 40 40 0 0 1 ${yEndX} ${yEndY}`} fill="none" stroke="#C9821E" strokeWidth="10" />
-                <path d={`M ${yEndX} ${yEndY} A 40 40 0 0 1 ${gEndX} ${gEndY}`} fill="none" stroke="#1C9A55" strokeWidth="10" />
-                {hasValue ? (
-                    <>
-                        <line x1="50" y1="50" x2={needleX} y2={needleY} stroke="currentColor" strokeWidth="1.6" strokeLinecap="round" className="text-om-ink" />
-                        <circle cx="50" cy="50" r="2.2" fill="currentColor" className="text-om-ink" />
-                    </>
-                ) : (
-                    <circle cx="50" cy="50" r="2.2" fill="currentColor" className="text-om-faintest" />
-                )}
-            </svg>
-            <div className="-mt-2 text-center leading-tight">
-                <div className={`font-mono font-medium ${oeeTextClass(hasValue ? p : null)}`} style={{ fontSize: size * 0.18 }}>
-                    {hasValue ? `${p.toFixed(1)}%` : 'N/A'}
-                </div>
-                <div className="font-mono text-om-faint uppercase tracking-[0.08em]" style={{ fontSize: size * 0.085 }}>
-                    OEE
-                </div>
-            </div>
-        </div>
-    );
 }
 
 function InboundQcOverview({ stats }) {
